Precision doesn't justify floating points so change to return an uint32_t. At 50 ms there is a +-20 Hz flicker

+++ Freqy/freqy.h
@@ -8,7 +8,7 @@
 #define FREQ_H
 
 void init_freqy();
-double get_input_freqy();
+uint32_t get_input_freqy();
 void set_output_freqy(float newFreqency);
 
 #endif
